+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add content and style to index ([#73])
 - use celery as queue to run background tasks ([#72])
 - implement web map service client ([#95])
+- add module containing the paper format specs ([#99])
 
 ### Bug Fixes
 
@@ -37,4 +38,5 @@
 [#72]: https://github.com/GIScience/sketch-map-tool/pull/72
 [#73]: https://github.com/GIScience/sketch-map-tool/pull/73
 [#95]: https://github.com/GIScience/sketch-map-tool/pull/95
+[#99]: https://github.com/GIScience/sketch-map-tool/pull/99
 [#102]: https://github.com/GIScience/sketch-map-tool/pull/102